% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_border.R
\name{add_border}
\alias{add_border}
\title{Función para añadir bordes a ggobjects para pruebas.}
\usage{
add_border(grob, color = "red")
}
\description{
Helper para agregar borde de depuración
}
\keyword{internal}
