% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fichatec.R
\name{fichatec}
\alias{fichatec}
\title{Genera una tabla de ficha técnica, para PowerPoint}
\usage{
fichatec(info, height = NULL)
}
\arguments{
\item{info}{Texto que irá en la tabla, en forma de objeto tribble}

\item{height}{Altura mínima de las filas (0.7 por defecto)}
}
\value{
Una tabla con flextable
}
\description{
Genera una tabla de ficha técnica, para PowerPoint
}
