% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspider.R
\name{ggspider}
\alias{ggspider}
\title{Spider/radar chart con ggplot}
\usage{
ggspider(
  data,
  labels_width = 30,
  show_axis_labels = TRUE,
  area_fill = FALSE,
  central_distance = 0,
  axis_name_offset = 0.2,
  ticks = 4,
  colores = c("#002060", "#ED7D31", "#FFC000"),
  min_value = NULL,
  max_value = NULL
)
}
\arguments{
\item{data}{Tabla para plotear. Primera columna `group` contiene nombres de las categorías de los grupos. Resto de columnas contienen los valores por cada variable.}

\item{area_fill}{Define si rellenar el área del grupo (FALSE por defecto).}

\item{central_distance}{Define distancia entre el centro y el origen del eje.}

\item{axis_name_offset}{Define cuánto se alejan los nombres de las variables del borde.}

\item{ticks}{Define cantidad de valores del eje a graficar.}

\item{colores}{Define paleta de colores.}

\item{min_value}{Define valor mínimo esperado para los ejes (si no se define, se calcula).}

\item{max_value}{Define valor máximo esperado para los ejes (si no se define, se calcula).}
}
\value{
Un objeto ggplot que puede ser exportado como dml o png.
}
\description{
Genera un gráfico de tipo radar o araña con valores absolutos. A partir de https://towardsdatascience.com/how-to-make-a-spider-chart-in-r-using-ggplot2-85a4f1898cab/
}
\keyword{internal}
