% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_escala.R
\name{plot_escala}
\alias{plot_escala}
\title{Genera un gráfico de barras horizontales apiladas para preguntas escalares, para PowerPoint}
\usage{
plot_escala(
  data,
  vars,
  levels = NULL,
  T2B = TRUE,
  show_notes = TRUE,
  show_n = TRUE,
  unit = NULL,
  unit_extra = TRUE
)
}
\arguments{
\item{data}{Un data frame con los datos en formato sav.}

\item{vars}{Variable(s) para plotear.}

\item{levels}{Niveles de la escala para ordenar la presentación. Va de negativo a positivo en cuatro opciones y después de eso, valores NA.}

\item{T2B}{Define si se calcula e incluye en el gráfico el "Top 2 Box" (TRUE por defecto).}

\item{unit}{Unidad de observación para declarar el N de la base usada para el gráfico.}

\item{unit_extra}{Define si añade una descripción adicional a la unidad de observación en la nota sobre el N de la base (TRUE por defecto). Requiere añadir el texto adicional desde params.}
}
\value{
Un objeto ggplot que puede ser exportado como dml.
}
\description{
Genera un gráfico de barras horizontales apiladas para preguntas escalares, para PowerPoint
}
