% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_escala_2.R
\name{plot_escala_2}
\alias{plot_escala_2}
\title{Genera un gráfico de barras horizontales apiladas para preguntas escalares, para PowerPoint}
\usage{
plot_escala_2(
  data = NULL,
  vars,
  ancho = 0.7,
  T2B = TRUE,
  title = NULL,
  unit_extra = TRUE,
  show_n = TRUE,
  show_notes = TRUE,
  x_labels = TRUE
)
}
\arguments{
\item{data}{Un data frame con los datos en formato sav.}

\item{vars}{Variable(s) para plotear.}

\item{T2B}{Define si se calcula e incluye en el gráfico el "Top 2 Box" (TRUE por defecto).}

\item{unit_extra}{Define si añade una descripción adicional a la unidad de observación en la nota sobre el N de la base (TRUE por defecto). Requiere añadir el texto adicional desde params.}

\item{levels}{Niveles de la escala para ordenar la presentación. Va de negativo a positivo en cuatro opciones y después de eso, valores NA.}
}
\value{
Un objeto ggplot que puede ser exportado como dml.
}
\description{
Genera un gráfico de barras horizontales apiladas para preguntas escalares, para PowerPoint
}
