% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_escala_gruppa.R
\name{plot_escala_gruppa}
\alias{plot_escala_gruppa}
\title{Barras horizontales apiladas para preguntas escalares y comparativas entre grupos}
\usage{
plot_escala_gruppa(vars, T2B = TRUE)
}
\arguments{
\item{vars}{Variable(s) para plotear.}

\item{T2B}{Define si se calcula e incluye en el gráfico el "Top 2 Box" (TRUE por defecto).}
}
\value{
Un objeto ggplot que puede ser exportado como dml.
}
\description{
Genera un gráfico de barras horizontales apiladas para preguntas escalares y comparativas entre grupos, para PowerPoint. Requiere distintas bases para cada grupo definido como lista en `vars`.
}
