% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gruppa_2.R
\name{plot_gruppa_2}
\alias{plot_gruppa_2}
\title{Barras horizontales apiladas para preguntas escalares y comparativas entre grupos}
\usage{
plot_gruppa_2(
  data = NULL,
  vars,
  ancho = 0.7,
  T2B = TRUE,
  title = NULL,
  unit_extra = TRUE,
  show_n = FALSE,
  show_notes = TRUE,
  x_labels = TRUE
)
}
\arguments{
\item{data}{Falta descripción.}

\item{vars}{Variable(s) para plotear.}

\item{ancho}{Falta descripción.}

\item{T2B}{Define si se calcula e incluye en el gráfico el "Top 2 Box" (TRUE por defecto).}

\item{title}{Falta descripción.}

\item{unit_extra}{Define si añade una descripción adicional a la unidad de observación en la nota sobre el N de la base (TRUE por defecto). Requiere añadir el texto adicional desde params.}

\item{show_n}{Falta descripción.}

\item{show_notes}{Falta descripción.}

\item{x_labels}{Falta descripción.}
}
\value{
Un objeto ggplot que puede ser exportado como dml.
}
\description{
Genera un gráfico de barras horizontales apiladas para preguntas escalares y comparativas entre grupos, para PowerPoint. Requiere distintas bases para cada grupo definido como lista en `vars`.
}
