% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar.R
\name{plot_radar}
\alias{plot_radar}
\title{Gráfico de radar para respuestas de Top Two Box}
\usage{
plot_radar(
  data = NULL,
  vars,
  labels_width = 30,
  show_axis_labels = FALSE,
  area_fill = FALSE,
  central_distance = 0,
  axis_name_offset = 0.2,
  ticks = 4,
  colores = c("#002060", "#ED7D31", "#FFC000"),
  min_value = 60,
  max_value = 100,
  title = NULL,
  unit_extra = TRUE
)
}
\arguments{
\item{data}{Un data frame con los datos en formato sav.}

\item{vars}{Variable o variables para plotear.}

\item{title}{Recibe un título para el gráfico. Si es NULL (opción por defecto), toma la etiqueta de la variable en vars.}

\item{unit_extra}{Define si añade una descripción adicional a la unidad de observación en la nota sobre el N de la base (TRUE por defecto). Requiere añadir el texto adicional desde params.}

\item{unit}{Unidad de observación para declarar el N de la base usada para el gráfico. Si es NULL (opción por defecto), toma el parámeto correspondiente al dataset en data ("data_unit") indicado en los params del yaml. Puede recibir un elemento directamente (no recomendado).}

\item{show_notes}{Define si se mostrarán las notas de pie de "N" y el mensaje sobre el redondeo (si TRUE). FALSE por defecto.}

\item{show_n}{Define si se mostrará el N como subtítulo (si TRUE). FALSE por defecto.}
}
\value{
Un objeto ggplot que puede ser exportado como dml o png.
}
\description{
Gráfico de radar para respuestas de Top Two Box
}
