% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_torta.R
\name{plot_torta}
\alias{plot_torta}
\title{Gráfico de torta para preguntas dicotómicas, para PowerPoint}
\usage{
plot_torta(
  data,
  var,
  title = NULL,
  title.pos = "top",
  legend.pos = "bottom",
  show_n = FALSE
)
}
\arguments{
\item{data}{Un data frame con los datos en formato sav.}

\item{var}{Variable para plotear.}

\item{title}{Recibe un título para el gráfico. Si es NULL (opción por defecto), toma la etiqueta de la variable en var.}

\item{title.pos}{Define la posición del título con respecto al gráfico. Si es 'top' (opción por defecto), va encima; si es 'left', el título va a la izquierda y el gráfico a la derecha.}

\item{show_n}{FALSE por defecto. Si es TRUE, calcula e incluye en el gráfico el "N" de registros que respondieron.}
}
\value{
Un objeto ggplot que puede ser exportado como dml.
}
\description{
Gráfico de torta para preguntas dicotómicas, para PowerPoint
}
