% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppng.R
\name{ppng}
\alias{ppng}
\title{Función para imprimir plot y elementos en word.}
\usage{
ppng(ggobj, fig.width = 6, fig.height = NULL)
}
\arguments{
\item{ggobj}{Objeto ggplot.}

\item{fig.width}{description}

\item{fig.height}{description}
}
\description{
Función para imprimir plot y elementos en word.
}
