% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_n.R
\name{show_n}
\alias{show_n}
\title{Muestra N de respuestas y nota al pie de p치gina}
\usage{
show_n(data, var, unit = NULL)
}
\arguments{
\item{data}{Un data frame con los datos en formato sav.}

\item{var}{Variable o variables a considerar para el c치lculo de respuestas.}

\item{unit}{Unidad de observaci칩n para declarar el N.}
}
\description{
Muestra N de respuestas y nota al pie de p치gina
}
