% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_n_2.R
\name{show_n_2}
\alias{show_n_2}
\title{Muestra N de respuestas y nota al pie de página}
\usage{
show_n_2(data, ..., vars = NULL)
}
\arguments{
\item{data}{Un data frame con los datos en formato sav.}

\item{vars}{Variable o variables a considerar para el cálculo de respuestas.}
}
\description{
Muestra N de respuestas y nota al pie de página
}
