% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{tribble}
\alias{select}
\alias{filter}
\alias{fp_border}
\alias{ph_location}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{officer}{\code{\link[officer]{fp_border}}, \code{\link[officer]{ph_location}}}

  \item{tibble}{\code{\link[tibble]{tribble}}}
}}

