% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_n.R
\name{show_n}
\alias{show_n}
\title{Muestra N de respuestas y nota al pie de página}
\usage{
show_n(data, ..., vars = NULL)
}
\arguments{
\item{data}{Un data frame con los datos en formato sav.}

\item{var}{Variable o variables a considerar para el cálculo de respuestas.}

\item{unit}{Unidad de observación para declarar el N.}

\item{unit_extra}{Define si añade una descripción adicional a la unidad de observación en la nota sobre el N de la base (TRUE por defecto). Requiere añadir el texto adicional desde params.}
}
\description{
Muestra N de respuestas y nota al pie de página
}
