% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_t2b.R
\name{tab_t2b}
\alias{tab_t2b}
\title{Tabla para respuestas de Top Two Box}
\usage{
tab_t2b(data = NULL, vars, labels_width = 30, unit_extra = FALSE)
}
\arguments{
\item{data}{Un data frame con los datos en formato sav.}

\item{vars}{Variable o variables para plotear.}

\item{labels_width}{description}

\item{unit_extra}{Define si a침ade una descripci칩n adicional a la unidad de observaci칩n en la nota sobre el N de la base (TRUE por defecto). Requiere a침adir el texto adicional desde params.}
}
\value{
Una tabla que puede ser exportada.
}
\description{
Tabla para respuestas de Top Two Box
}
